# use sync.Pool to reduce gc presure on hot path