--- a/ast/flag_test.go
+++ b/ast/flag_test.go
@@ -33,7 +33,7 @@ type testFlagSuite struct {
 }
 
 func (ts *testFlagSuite) SetUpSuite(c *C) {
-	ts.Parser = parser.New()
+	ts.Parser = parser.NewParser()
 }
 
 func (ts *testFlagSuite) TestHasAggFlag(c *C) {