--- a/ast/format_test.go
+++ b/ast/format_test.go
@@ -84,7 +84,7 @@ func (ts *testAstFormatSuite) TestAstFormat(c *C) {
 	for _, tt := range testcases {
 		expr := fmt.Sprintf("select %s", tt.input)
 		charset, collation := getDefaultCharsetAndCollate()
-		stmts, err := parser.New().Parse(expr, charset, collation)
+		stmts, err := parser.NewParser().Parse(expr, charset, collation)
 		node := stmts[0].(*ast.SelectStmt).Fields.Fields[0].Expr
 		c.Assert(err, IsNil)