--- a/expression/aggregation/agg_to_pb_test.go
+++ b/expression/aggregation/agg_to_pb_test.go
@@ -106,7 +106,7 @@ type testEvaluatorSuite struct {
 }
 
 func (s *testEvaluatorSuite) SetUpSuite(c *C) {
-	s.Parser = parser.New()
+	s.Parser = parser.NewParser()
 	s.ctx = mock.NewContext()
 }