--- a/parser/bench_test.go
+++ b/parser/bench_test.go
@@ -22,9 +22,11 @@ func BenchmarkSysbenchSelect(b *testing.B) {
 	sql := "SELECT pad FROM sbtest1 WHERE id=1;"
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_, err := parser.Parse(sql, "", "")
-		if err != nil {
-			b.Fatal(err)
+		for j := 0; j < 1024; j++ {
+			_, err := parser.Parse(sql, "", "")
+			if err != nil {
+				b.Fatal(err)
+			}
 		}
 	}
 	b.ReportAllocs()
@@ -39,10 +41,12 @@ func BenchmarkParse(b *testing.B) {
 	parser := New()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		for _, v := range table {
-			_, err := parser.Parse(v, "", "")
-			if err != nil {
-				b.Failed()
+		for j := 0; j < 1024; j++ {
+			for _, v := range table {
+				_, err := parser.Parse(v, "", "")
+				if err != nil {
+					b.Fatal(err)
+				}
 			}
 		}
 	}