--- a/parser/yy_parser.go
+++ b/parser/yy_parser.go
@@ -185,11 +185,14 @@ func toInt(l yyLexer, lval *yySymType, str string) int {
 	default:
 		lval.item = n
 	}
+	if d, ok := lval.item.(*types.MyDecimal); ok {
+		d.Free()
+	}
 	return intLit
 }
 
 func toDecimal(l yyLexer, lval *yySymType, str string) int {
-	dec := new(types.MyDecimal)
+	dec := types.NewMyDecimal()
 	err := dec.FromString(hack.Slice(str))
 	if err != nil {
 		l.Errorf("decimal literal: %v", err)