--- a/tidb.go
+++ b/tidb.go
@@ -136,7 +136,7 @@ func SetCommitRetryLimit(limit int) {
 func Parse(ctx context.Context, src string) ([]ast.StmtNode, error) {
 	log.Debug("compiling", src)
 	charset, collation := ctx.GetSessionVars().GetCharsetInfo()
-	p := parser.New()
+	p := parser.NewParser()
 	p.SetSQLMode(ctx.GetSessionVars().SQLMode)
 	stmts, err := p.Parse(src, charset, collation)
 	if err != nil {