Below is the benchmark, I am not sure why the ~~~memory usage does not decrease.~~~ (we just reuse the object. Memory allocation is still original. Using sync.Pool can only reduce gc pressure. )
I am inverstigating this. In additon, I will add different workload benchmark test later.
```
name              old time/op    new time/op    delta
SysbenchSelect-8    46.8ms ± 1%    46.5ms ± 5%  -0.55%  (p=0.044 n=18+20)
Parse-8              147ms ± 3%     145ms ± 2%  -1.61%  (p=0.000 n=20+17)

name              old alloc/op   new alloc/op   delta
SysbenchSelect-8    18.4MB ± 0%    18.4MB ± 0%    ~     (p=0.057 n=17+18)
Parse-8             51.7MB ± 0%    51.7MB ± 0%    ~     (p=0.072 n=19+20)

name              old allocs/op  new allocs/op  delta
SysbenchSelect-8      184k ± 0%      184k ± 0%    ~     (all equal)
Parse-8               707k ± 0%      707k ± 0%    ~     (all equal)
```