@zhexuany 2018-01-28
@coocood Here is another example using `sync.Pool` can benefit on reducing cpu usage.