@coocood 2018-01-28
@zhexuany That's because the `toDecimal` function is never called.