@coocood 2018-01-28
@zhexuany 
It doesn't make sense that reducing one small allocation improves noticeable performance if there are hundreds of allocations already.