@zhexuany 2018-01-28
Here is benchmark result for different workloads. 
```
SysbenchSelect1024-8    4.74ms ±10%    4.62ms ± 8%  -2.53%  (p=0.000 n=39+36)
SysbenchSelect256-8     1.21ms ±13%    1.20ms ±16%    ~     (p=0.307 n=40+40)
SysbenchSelect128-8      644µs ±11%     611µs ±13%  -5.16%  (p=0.000 n=38+39)
ParseInt4096-8          82.2ms ±11%    85.2ms ±18%  +3.70%  (p=0.012 n=39+40)
ParseInt1024-8          21.6ms ±12%    22.3ms ±13%    ~     (p=0.063 n=39+39)
ParseInt512-8           10.6ms ±14%    11.1ms ±15%  +4.72%  (p=0.000 n=40+39)
ParseInt256-8           5.04ms ± 9%    5.35ms ± 7%  +6.21%  (p=0.000 n=39+40)

name                  old alloc/op   new alloc/op   delta
SysbenchSelect1024-8    1.84MB ± 0%    1.84MB ± 0%  +0.00%  (p=0.022 n=32+37)
SysbenchSelect256-8      459kB ± 0%     459kB ± 0%    ~     (all equal)
SysbenchSelect128-8      229kB ± 0%     229kB ± 0%    ~     (all equal)
ParseInt4096-8          28.4MB ± 0%    28.4MB ± 0%  +0.01%  (p=0.000 n=39+24)
ParseInt1024-8          7.09MB ± 0%    7.10MB ± 0%  +0.02%  (p=0.000 n=40+40)
ParseInt512-8           3.55MB ± 0%    3.55MB ± 0%  +0.04%  (p=0.000 n=39+40)
ParseInt256-8           1.77MB ± 0%    1.77MB ± 0%  +0.04%  (p=0.000 n=40+40)

name                  old allocs/op  new allocs/op  delta
SysbenchSelect1024-8     18.4k ± 0%     18.4k ± 0%    ~     (all equal)
SysbenchSelect256-8      4.61k ± 0%     4.61k ± 0%    ~     (all equal)
SysbenchSelect128-8      2.30k ± 0%     2.30k ± 0%    ~     (all equal)
ParseInt4096-8            360k ± 0%      360k ± 0%  +0.00%  (p=0.000 n=40+40)
ParseInt1024-8           90.1k ± 0%     90.1k ± 0%  +0.00%  (p=0.000 n=40+40)
ParseInt512-8            45.1k ± 0%     45.1k ± 0%  +0.00%  (p=0.000 n=40+40)
ParseInt256-8            22.5k ± 0%     22.5k ± 0%  +0.00%  (p=0.000 n=40+40
```