@ngaut 2018-01-31
The benchmark results shows that it is not worth to use sync.pool.