@ngaut 2018-01-31
I will close this PR. please feel free to reopen it if you have different thoughts.